Added correct icon; moved Client and Animal components and services to correct folder and updated references
